requirements.txt
pytest==8.2.1
playwright
pytest-playwright==0.5.0
python-dotenv==1.0.1
pytest-xdist==3.6.1
pytest-html==4.1.1
